requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django>=4.2.0
djangorestframework==3.16.0
idna==3.10
lxml==5.3.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
python-docx==1.1.2
requests>=2.31.0
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openai>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
netlify-python>=0.1.0
google-generativeai>=0.3.0